services/trade/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
greenlet==3.0.3
httpx==0.27.0

# Async DB support
alembic==1.13.1
SQLAlchemy==2.0.29
asyncpg>=0.29,<1.0

# Optional: Only needed if you use sync connections somewhere
psycopg2-binary==2.9.9

# Pydantic for config and validation
pydantic[email]==2.6.4
pydantic-settings>=2.0,<3.0

# .env loader
python-dotenv==1.0.1

# If you use JWT or token logic in the future (optional)
# python-jose[cryptography]==3.3.0
